requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic[email]==2.5.0
pydantic-settings==2.1.0
jinja2==3.1.2
email-validator==2.1.0
psycopg2-binary==2.9.9
alembic==1.12.
bcrypt==4.1.2
Pillow==10.1.0
pytest
pytest-asyncio
httpx
